requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
setuptools>=65.0.0
pyspark>=3.4.0
fastapi>=0.104.0
